dt-bindings: fix compilation error of the example in intel,lgm-emmc-phy.yaml
Running:
export DT_SCHEMA_FILES=Documentation/devicetree/bindings/phy/intel,lgm-emmc-phy.yaml
'make dt_binding_check'

gives a compilation error. This is because in the example there
is the label 'emmc-phy' but labels are not allowed to have '-' sing.
Replace the '-' with '_' to fix the error.

Fixes: 5bc999108025 ("dt-bindings: phy: intel-emmc-phy: Add YAML
schema for LGM eMMC PHY")
Acked-by: Rob Herring <robh@kernel.org>
Signed-off-by: Dafna Hirschfeld <dafna.hirschfeld@collabora.com>
Signed-off-by: Kishon Vijay Abraham I <kishon@ti.com>